Create What I have so far
All this succeeds in doing is making to servo move in somewhat direct relation to noise level. The movement of the servo is fidgety, but not in an outrageous way. I am looking to make the movement of the servo no less responsive, but filter out sudden, small changes in sound and have it be more responsive to the bigger changes in sound. 
On top of this, using an avg. of the number of readings it takes, I would like it to change its baseline "energy" level, so to speak, so that in a louder room the sculpture would be more engaged, but not necessarily as sensitive to the noises around it.